revert(storage): defer renderer migration cleanup
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
