+++ app/utils/storageKeys.test.ts
@@ -8,7 +8,6 @@ import {
   storageRemove,
   storageSet,
   storageSetJSON,
-  storageUpdate,
 } from './storageKeys';
 
 describe('storageKeys', () => {
@@ -37,10 +36,6 @@ describe('storageKeys', () => {
 
   it('prefixes keys with opencode', () => {
     expect(storageKey('foo')).toBe('opencode.foo');
-    expect(storageKey(StorageKeys.settings.modelVisibility)).toBe('opencode.global.dat:model');
-    expect(storageKey(StorageKeys.settings.disabledModels)).toBe(
-      'opencode.settings.disabledModels.v1',
-    );
   });
 
   it('round-trips strings via storageSet and storageGet', () => {
@@ -90,7 +85,6 @@ describe('storageKeys', () => {
           removeItem: vi.fn((key: string) => {
             delete electronStore[key];
           }),
-          migrate: vi.fn(() => true),
         },
       },
     });
@@ -109,7 +103,6 @@ describe('storageKeys', () => {
           getItem: vi.fn(() => null),
           setItem: vi.fn(() => false),
           removeItem: vi.fn(() => false),
-          migrate: vi.fn(() => true),
         },
       },
     });
@@ -121,121 +114,6 @@ describe('storageKeys', () => {
     // Then: both false acknowledgements remain observable to rollback callers.
     expect(setResult).toBe(false);
     expect(removeResult).toBe(false);
-  });
-
-  it('prefixes and forwards one atomic Electron storage update', () => {
-    // Given: Electron exposes one observable bundle-update boundary.
-    const update = vi.fn(() => true);
-    vi.stubGlobal('window', {
-      localStorage: window.localStorage,
-      electronAPI: {
-        persistentStorage: {
-          getItem: vi.fn(() => null),
-          setItem: vi.fn(() => true),
-          removeItem: vi.fn(() => true),
-          migrate: vi.fn(() => true),
-          update,
-        },
-      },
-    });
-
-    // When: credential metadata, secret, and removal are submitted together.
-    const result = storageUpdate({
-      [StorageKeys.auth.serverUrl]: 'https://new.example',
-      [StorageKeys.auth.credentials]: 'new-secret',
-      [StorageKeys.auth.codexBridgeToken]: null,
-    });
-
-    // Then: one prefixed payload crosses the Electron boundary unchanged.
-    expect(result).toBe(true);
-    expect(update).toHaveBeenCalledWith({
-      'opencode.auth.serverUrl.v1': 'https://new.example',
-      'opencode.auth.credentials.v1': 'new-secret',
-      'opencode.auth.codexBridgeToken.v1': null,
-    });
-  });
-
-  it('throttles failed migration retries while keeping legacy storage readable', async () => {
-    // Given: Electron rejects the first atomic migration of legacy renderer state.
-    vi.resetModules();
-    const now = vi.spyOn(Date, 'now').mockReturnValue(1_000);
-    const key = 'opencode.settings.textTransformers.v1';
-    const migrate = vi.fn(() => false);
-    const localStorage = {
-      length: 1,
-      key: vi.fn(() => key),
-      getItem: vi.fn((requestedKey: string) => (requestedKey === key ? 'legacy-snippets' : null)),
-      setItem: vi.fn(),
-      removeItem: vi.fn(),
-    };
-    vi.stubGlobal('window', {
-      localStorage,
-      electronAPI: {
-        persistentStorage: {
-          getItem: vi.fn(() => null),
-          setItem: vi.fn(() => true),
-          removeItem: vi.fn(() => true),
-          migrate,
-        },
-      },
-    });
-    const freshStorage = await import('./storageKeys');
-
-    // When: callers read repeatedly inside one retry window and once after it advances.
-    const first = freshStorage.storageGet(freshStorage.StorageKeys.settings.textTransformers);
-    const second = freshStorage.storageGet(freshStorage.StorageKeys.settings.textTransformers);
-    now.mockReturnValue(2_000);
-    const third = freshStorage.storageGet(freshStorage.StorageKeys.settings.textTransformers);
-
-    // Then: every read uses legacy data but only one migration runs per retry window.
-    expect(first).toBe('legacy-snippets');
-    expect(second).toBe('legacy-snippets');
-    expect(third).toBe('legacy-snippets');
-    expect(migrate).toHaveBeenCalledTimes(2);
-    expect(migrate).toHaveBeenLastCalledWith({ [key]: 'legacy-snippets' });
-  });
-
-  it('uses Electron storage when the localStorage getter is unavailable', async () => {
-    // Given: a sandbox policy makes localStorage access throw while Electron storage remains healthy.
-    vi.resetModules();
-    const key = 'opencode.settings.enterToSend.v1';
-    const electronStore: Record<string, string> = { [key]: 'electron-value' };
-    const migrate = vi.fn(() => true);
-    const windowStub = {
-      electronAPI: {
-        persistentStorage: {
-          getItem: vi.fn((requestedKey: string) => electronStore[requestedKey] ?? null),
-          setItem: vi.fn((requestedKey: string, value: string) => {
-            electronStore[requestedKey] = value;
-            return true;
-          }),
-          removeItem: vi.fn((requestedKey: string) => {
-            delete electronStore[requestedKey];
-            return true;
-          }),
-          migrate,
-        },
-      },
-    };
-    Object.defineProperty(windowStub, 'localStorage', {
-      get() {
-        throw new DOMException('Access denied', 'SecurityError');
-      },
-    });
-    vi.stubGlobal('window', windowStub);
-    const freshStorage = await import('./storageKeys');
-
-    // When: callers read, write, and remove through the generic storage helpers.
-    const initial = freshStorage.storageGet(freshStorage.StorageKeys.settings.enterToSend);
-    const saved = freshStorage.storageSet(freshStorage.StorageKeys.settings.enterToSend, 'updated');
-    const removed = freshStorage.storageRemove(freshStorage.StorageKeys.settings.enterToSend);
-
-    // Then: no helper throws, Electron owns every operation, and migration is not attempted.
-    expect(initial).toBe('electron-value');
-    expect(saved).toBe(true);
-    expect(removed).toBe(true);
-    expect(electronStore[key]).toBeUndefined();
-    expect(migrate).not.toHaveBeenCalled();
   });
 
   it('exposes codexActiveThread key for codex session persistence', () => {
